# documentation/api.spec.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Music Taste Explorer
  description: >-
    "Create playlists to explore users's musical taste differences and similarities."
  termsOfService: adambechtold.xyz
  contact:
    name: Adam Bechtold
servers:
  - url: http://localhost:4000
tags:
  - name: users
    description: Operations related to users.
  - name: music
    description: Operations related to music.
  - name: authentication
    description: Operations related to authentication.
paths:
  /api/users/{userId}:
    get:
      summary: Get User by Id
      description: Get a user by their id
      operationId: getUserById
      tags:
        - users
      parameters:
        - name: userId
          in: path
          description: The id of the user to retrieve
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A user object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/music/play-track/{trackId}:
    put:
      summary: Play the Given Track
      description: Plays the provided track. This will always affect user 1 (per-user control coming soon).
      operationId: playTrack
      tags:
        - music
      parameters:
        - name: trackId
          in: path
          description: The id of the track to play
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Track played successfully
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Track not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/music/play-tracks:
    put:
      summary: Play the Given Tracks
      description: Plays the provided tracks. This will always affect user 1 (per-user control coming soon).
      operationId: playTracks
      tags:
        - music
      requestBody:
        description: The tracks to play
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                trackIds:
                  type: array
                  items:
                    type: integer
                offset:
                  type: integer
                  description: The offset to start playing the tracks at.
                  default: 0
      responses:
        "204":
          description: Tracks played successfully
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Track not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/music/transfer-playback:
    put:
      summary: Transfer Playback
      description: Transfer playback to the given device
      operationId: transferPlayback
      tags:
        - music
      requestBody:
        description: The device to transfer playback to
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId:
                  type: string
      responses:
        "204":
          description: Playback transferred successfully
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/music/playlists:
    get:
      summary: Get Playlist
      description: Get a playlist based on the users and the preference type
      operationId: getPlaylist
      tags:
        - music
      parameters:
        - name: userId1
          in: query
          description: The id of the first user
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
        - name: userId2
          in: query
          description: The id of the second user
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
        - name: preferenceType
          in: query
          description: The type of preference to use when creating the playlist
          required: true
          schema:
            type: string
            enum:
              - "USER1-ONLY"
              - "USER2-ONLY"
              - "BOTH"
      responses:
        "200":
          description: A list of tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /api/music/tracks:
    get:
      summary: Get Tracks
      description: Get Tracks that fit the query
      operationId: getTracks
      tags:
        - music
      parameters:
        - name: name
          in: query
          description: The name of the track
          required: false
          schema:
            type: string
        - name: artist
          in: query
          description: The name of the artist
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Track(s) that fit the query
          content:
            application/json:
              schema:
                type: object
                properties:
                  tracks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Track"
        "400":
          description: Invalid Query Parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Track(s) not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/music/lastfm-listens/{id}/track:
    get:
      summary: Get the Track from last.fm Listen
      description: Get the track information the last.fm listen
      operationId: getTrackFromLastfmListen
      tags:
        - music
      parameters:
        - name: id
          in: path
          description: The id of the last.fm listen
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A track object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: last.fm Listen not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/spotify/token:
    get:
      summary: Get Spotify Access Token
      description: Get a Spotify access token based on the currently signed in user.
      operationId: getSpotifyAccessToken
      tags:
        - music
        - authentication
      responses:
        "200":
          description: A Spotify access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        "401":
          description: User is not authorized or no user is signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    User:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
        lastfmAccount:
          $ref: "#/components/schemas/LastfmAccount"
    LastfmAccount:
      type: object
      required:
        - username
        - registeredAt
        - url
        - playCount
        - trackCount
      properties:
        username:
          type: string
        registeredAt:
          type: string
          format: date-time
        url:
          type: string
          format: uri
        playCount:
          type: integer
        trackCount:
          type: integer
    Playlist:
      type: object
      required:
        - tracks
      properties:
        tracks:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Track"
    Track:
      type: object
      required:
        - name
        - artists
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          format: uri
        artists:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
    UserId:
      type: integer
      minimum: 1
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    AccessToken:
      type: object
      required:
        - accessToken
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        service:
          type: string
          enum: ["SPOTIFY"]